Move some utils around
