+++ packages/client/src/generation/getDMMF.ts
@@ -2,8 +2,8 @@ import type { DMMF } from '@prisma/generator-helper'
 import type { GetDMMFOptions } from '@prisma/internals'
 import { getDMMF as getRawDMMF } from '@prisma/internals'
 
-import type { DMMF as PrismaClientDMMF } from '../runtime/dmmf-types'
 import { externalToInternalDmmf } from '../runtime/externalToInternalDmmf'
+import type { DMMF as PrismaClientDMMF } from './dmmf-types'
 
 export function getPrismaClientDMMF(dmmf: DMMF.Document): PrismaClientDMMF.Document {
   return externalToInternalDmmf(dmmf)
