+++ packages/client/src/generation/utils/argIsInputType.ts
@@ -1,0 +1,9 @@
+import type { DMMF } from '../dmmf-types'
+
+export function argIsInputType(arg: DMMF.ArgType): arg is DMMF.InputType {
+  if (typeof arg === 'string') {
+    return false
+  }
+
+  return true
+}
