+++ packages/client/src/runtime/index.ts
@@ -4,6 +4,8 @@ import * as Types from './core/types'
 import { Payload } from './core/types'
 import type { PrismaPromise } from './core/types/Public'
 
+export { DMMFHelper as DMMFClass } from '../generation/dmmf'
+export { type BaseDMMF, DMMF } from '../generation/dmmf-types'
 export { PrismaClientInitializationError } from './core/errors/PrismaClientInitializationError'
 export { PrismaClientKnownRequestError } from './core/errors/PrismaClientKnownRequestError'
 export { PrismaClientRustPanicError } from './core/errors/PrismaClientRustPanicError'
@@ -18,8 +20,6 @@ export {
 } from './core/metrics/MetricsClient'
 export type { FieldRef } from './core/model/FieldRef'
 export { defineDmmfProperty } from './core/runtimeDataModel'
-export { DMMFHelper as DMMFClass } from './dmmf'
-export { type BaseDMMF, DMMF } from './dmmf-types'
 export type { PrismaClientOptions } from './getPrismaClient'
 export { getPrismaClient } from './getPrismaClient'
 export { objectEnumValues } from './object-enums'
