+++ packages/client/src/utils/lowerCase.ts
@@ -1,0 +1,7 @@
+/**
+ * Converts the first character of a word to lower case
+ * @param name
+ */
+export function lowerCase(name: string): string {
+  return name.substring(0, 1).toLowerCase() + name.substring(1)
+}
