+++ packages/internals/src/index.ts
@@ -48,7 +48,7 @@ export { sendPanic } from './sendPanic'
 export * from './tracing/types'
 export type { DatabaseCredentials } from './types'
 export { assertNever } from './utils/assertNever'
-export { byline } from './utils/byline'
+export { default as byline } from './utils/byline'
 export { callOnceOnSuccess } from './utils/callOnce'
 export { canPrompt } from './utils/canPrompt'
 export { drawBox } from './utils/drawBox'
